Seperated out FCleverTapProperties related JNI wrappers
