+++ Plugins/CleverTap/Source/CleverTap/Private/Android/AndroidCleverTapPropertiesJNI.h
@@ -1,0 +1,15 @@
+// Copyright CleverTap All Rights Reserved.
+#pragma once
+
+#include "CleverTapProperties.h"
+
+#include "Android/AndroidApplication.h"
+
+namespace CleverTapSDK { namespace Android { namespace JNI {
+
+jobject ConvertCleverTapPropertiesToJavaMap(JNIEnv* Env, const FCleverTapProperties& Properties);
+jobject ConvertArrayOfCleverTapPropertiesToJavaArrayOfMap(JNIEnv* Env, const TArray<FCleverTapProperties>& Array);
+
+FCleverTapProperties ConvertJavaMapToCleverTapProperties(JNIEnv* Env, jobject JavaMap);
+
+}}} // namespace CleverTapSDK::Android::JNI
