+++ Plugins/CleverTap/Source/CleverTap/Private/Android/AndroidCleverTapSDK.cpp
@@ -2,6 +2,7 @@
 #include "Android/AndroidCleverTapSDK.h"
 
 #include "Android/AndroidCleverTapJNI.h"
+#include "Android/AndroidCleverTapPropertiesJNI.h"
 #include "Android/AndroidJNIUtilities.h"
 
 #include "CleverTapInstance.h"
